+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## New Features
 
+- Add scandir() as Storage plugin API to speedup large directory read when generating global index `PR #1340`
 - Declare support for Python 3.11 `PR #1338`
 - Move Docker to build in 3.11 `PR #1341`